test_fread: add fixtures and separate type/eq checks
This is now meant to check all 3 samples for types of main parts (and
hopefully other stuff given the fixture), but the third sample was
raising a DNS issue on my Python but not browser so it only does 2.

It also checks the equality of the SdmxJsonDataMessage instances
returned by fread_json from remote and local paths, but the __eq__()
method isn't implemented yet, so this is a known fail.