Fix test_fread for agri.json
There is a typo in agri.json (dataset instead of dataSet) which means
that the test_get_attributes was failing for that sample message. Now
it is manually fixed for the message read from local file path, and
that version is used for all the tests.

The test_fread_json_local_and_remote_eq() is disabled for agri.json
since I don't know how to fix the mistake in the remote before using
fread_json, and I would prefer to run the tests with a corrected
version.